pyproject.toml
[project]
name = "generative-ai-using-langchain"
version = "0.1.0"
description = "All my notes for the LangChain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.6",
    "huggingface-hub>=0.36.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.1.3",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.7",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "transformers>=4.57.5",
]
